- moves feedPreference to its own package - adds FeedPreferenceRepository and FeedPreferenceState
